requirements.txt
scrapy
pymongo
pymysql
psycopg2-binary 